+++ b2sdk/sync/policy.py
@@ -64,16 +64,16 @@ class AbstractFileSyncPolicy(metaclass=ABCMeta):
         AbstractSyncEncryptionSettingsProvider = SERVER_DEFAULT_SYNC_ENCRYPTION_SETTINGS_PROVIDER,
     ):
         """
-        :param b2sdk.v1.AbstractSyncPath source_path: source file object
-        :param b2sdk.v1.AbstractFolder source_folder: source folder object
-        :param b2sdk.v1.AbstractSyncPath dest_path: destination file object
-        :param b2sdk.v1.AbstractFolder dest_folder: destination folder object
+        :param b2sdk.v2.AbstractSyncPath source_path: source file object
+        :param b2sdk.v2.AbstractFolder source_folder: source folder object
+        :param b2sdk.v2.AbstractSyncPath dest_path: destination file object
+        :param b2sdk.v2.AbstractFolder dest_folder: destination folder object
         :param int now_millis: current time in milliseconds
         :param int keep_days: days to keep before delete
-        :param b2sdk.v1.NewerFileSyncMode newer_file_mode: setting which determines handling for destination files newer than on the source
+        :param b2sdk.v2.NewerFileSyncMode newer_file_mode: setting which determines handling for destination files newer than on the source
         :param int compare_threshold: when comparing with size or time for sync
-        :param b2sdk.v1.COMPARE_VERSION_MODES compare_version_mode: how to compare source and destination files
-        :param b2sdk.v1.AbstractSyncEncryptionSettingsProvider encryption_settings_provider: encryption setting provider
+        :param b2sdk.v2.CompareVersionMode compare_version_mode: how to compare source and destination files
+        :param b2sdk.v2.AbstractSyncEncryptionSettingsProvider encryption_settings_provider: encryption setting provider
         """
         self._source_path = source_path
         self._source_folder = source_folder
@@ -120,11 +120,11 @@ class AbstractFileSyncPolicy(metaclass=ABCMeta):
         Compare two files and determine if the the destination file
         should be replaced by the source file.
 
-        :param b2sdk.v1.AbstractSyncPath source_path: source file object
-        :param b2sdk.v1.AbstractSyncPath dest_path: destination file object
+        :param b2sdk.v2.AbstractSyncPath source_path: source file object
+        :param b2sdk.v2.AbstractSyncPath dest_path: destination file object
         :param int compare_threshold: compare threshold when comparing by time or size
-        :param b2sdk.v1.CompareVersionMode compare_version_mode: source file version comparator method
-        :param b2sdk.v1.NewerFileSyncMode newer_file_mode: newer destination handling method
+        :param b2sdk.v2.CompareVersionMode compare_version_mode: source file version comparator method
+        :param b2sdk.v2.NewerFileSyncMode newer_file_mode: newer destination handling method
         """
         # Optionally set a compare threshold for fuzzy comparison
         compare_threshold = compare_threshold or 0
@@ -372,7 +372,7 @@ def make_b2_delete_note(version, index, transferred):
     """
     Create a note message for delete action.
 
-    :param b2sdk.v1.FileVersionInfo version: an object which contains file version info
+    :param b2sdk.v2.FileVersion version: an object which contains file version info
     :param int index: file version index
     :param bool transferred: if True, file has been transferred, False otherwise
     """
@@ -393,9 +393,9 @@ def make_b2_delete_actions(
     """
     Create the actions to delete files stored on B2, which are not present locally.
 
-    :param b2sdk.v1.AbstractSyncPath source_path: source file object
-    :param b2sdk.v1.AbstractSyncPath dest_path: destination file object
-    :param b2sdk.v1.AbstractFolder dest_folder: destination folder
+    :param b2sdk.v2.AbstractSyncPath source_path: source file object
+    :param b2sdk.v2.AbstractSyncPath dest_path: destination file object
+    :param b2sdk.v2.AbstractFolder dest_folder: destination folder
     :param bool transferred: if True, file has been transferred, False otherwise
     """
     if dest_path is None:
@@ -431,9 +431,9 @@ def make_b2_keep_days_actions(
     only the 25-day old version can be deleted.  The 15 day-old version
     was visible 10 days ago.
 
-    :param b2sdk.v1.AbstractSyncPath source_path: source file object
-    :param b2sdk.v1.AbstractSyncPath dest_path: destination file object
-    :param b2sdk.v1.AbstractFolder dest_folder: destination folder object
+    :param b2sdk.v2.AbstractSyncPath source_path: source file object
+    :param b2sdk.v2.AbstractSyncPath dest_path: destination file object
+    :param b2sdk.v2.AbstractFolder dest_folder: destination folder object
     :param bool transferred: if True, file has been transferred, False otherwise
     :param int keep_days: how many days to keep a file
     :param int now_millis: current time in milliseconds