+++ b2sdk/v1/exception.py
@@ -8,7 +8,7 @@
 #
 ######################################################################
 
-from b2sdk._v2.exception import *  # noqa
+from b2sdk.v2.exception import *  # noqa
 v2DestFileNewer = DestFileNewer