+++ b2sdk/v1/file_version.py
@@ -12,7 +12,7 @@ from typing import Optional
 import datetime
 import functools
 
-from b2sdk import _v2 as v2
+from b2sdk import v2
 from . import api as v1api